+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ##
 ## Added
+- replace docxf with pdf as a form template
 - ar-SA, ca-ES, da-DK, eu-ES, fi-FI, gl-ES, he-IL, hu-HU, hy-AM, id-ID, ms-MY, nb-NO, ro-RO, si-LK, sl-SI, sq-AL, sr-Cyrl-RS, sr-Latn-RS, tr-TR, ur-PK and zh-TW empty file templates
 - support for plone 6.1
 - document formats submodule